pythoncode/requirements.txt
googlesearch-python
requests_html
beautifulsoup4
sys
time